use int for zhop

+++ src/gcodeExport.h
@@ -155,7 +155,7 @@ private:
     double currentSpeed;
     int zPos;
     bool isRetracted;
-    double isZHopped;
+    int isZHopped;
 
     double last_coasted_amount_mm3; //!< The coasted amount of filament to be primed on the first next extrusion. (same type as GCodeExport::extrusion_amount)
     double retractionPrimeSpeed;
